bta_gattc: Fix number of max gatt connection client control blocks
Each GATT client application in worst case might have instance per
connection.
Without this change and having LeAudio support which per set of CSIP
based devices might take up to 12 GATT clients (vc, leaudio, csip,
battery service, bass, hap) * 2, the pool will end very quickly

Bug: 325438660
Test: mmm packages/modules/Bluetooth
Flag: Exempt, just increase number of control blocks for avoid future
issues
Test: manual testing multiple devices connection

+++ system/bta/gatt/bta_gattc_utils.cc
@@ -98,7 +98,7 @@ tBTA_GATTC_CLCB* bta_gattc_find_clcb_by_cif(uint8_t client_if,
                                             const RawAddress& remote_bda,
                                             tBT_TRANSPORT transport) {
   tBTA_GATTC_CLCB* p_clcb = &bta_gattc_cb.clcb[0];
-  uint8_t i;
+  int i;
 
   for (i = 0; i < BTA_GATTC_CLCB_MAX; i++, p_clcb++) {
     if (p_clcb->in_use && p_clcb->p_rcb->client_if == client_if &&
@@ -118,7 +118,7 @@ tBTA_GATTC_CLCB* bta_gattc_find_clcb_by_cif(uint8_t client_if,
  ******************************************************************************/
 tBTA_GATTC_CLCB* bta_gattc_find_clcb_by_conn_id(uint16_t conn_id) {
   tBTA_GATTC_CLCB* p_clcb = &bta_gattc_cb.clcb[0];
-  uint8_t i;
+  int i;
 
   for (i = 0; i < BTA_GATTC_CLCB_MAX; i++, p_clcb++) {
     if (p_clcb->in_use && p_clcb->bta_conn_id == conn_id) return p_clcb;
@@ -138,7 +138,7 @@ tBTA_GATTC_CLCB* bta_gattc_find_clcb_by_conn_id(uint16_t conn_id) {
 tBTA_GATTC_CLCB* bta_gattc_clcb_alloc(tGATT_IF client_if,
                                       const RawAddress& remote_bda,
                                       tBT_TRANSPORT transport) {
-  uint8_t i_clcb = 0;
+  int i_clcb = 0;
   tBTA_GATTC_CLCB* p_clcb = NULL;
 
   for (i_clcb = 0; i_clcb < BTA_GATTC_CLCB_MAX; i_clcb++) {
